Add DataFrame.median()
